requirements.txt
asgiref==3.6.0
async-generator==1.10
attrs==23.1.0
Babel==2.12.1
beautifulsoup4==4.12.2
black @ git+https://github.com/psf/black@a4032dce645b83e1faccc7274864869ddfe279fc
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.6
comtypes==1.1.14
crispy-bootstrap4==2022.1
defusedxml==0.7.1
diff-match-patch==20230430
dj-database-url==2.0.0
Django==4.2
django-bootstrap4==23.1
django-crispy-forms==2.0
django-heroku==0.3.1
django-import-export==3.1.0
django-translation==1.2
et-xmlfile==1.1.0
exceptiongroup==1.1.1
execnet==1.9.0
Faker==12.0.1
gunicorn==20.1.0
h11==0.14.0
idna==3.4
iniconfig==2.0.0
MarkupPy==1.14
mixer==7.2.2
mypy==1.3.0
mypy-extensions==1.0.0
numpy==1.24.3
odfpy==1.4.1

openpyxl==3.1.1

outcome==1.2.0
packaging==23.1
pandas==2.0.1
pathspec==0.11.1
platformdirs==3.5.1
pluggy==1.0.0
psycopg2==2.9.6
pycparser==2.21
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
PySocks==1.7.1
pytest==7.3.1
pytest-django==4.5.2
pytest-xdist==3.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pyttsx3==2.90
pytz==2023.3
PyYAML==6.0
requests==2.26.0
selenium==4.9.1
simplejson==3.17.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
tablib==3.5.0
tomli==2.0.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
webdriver-manager==3.8.6
whitenoise==6.4.0
wsproto==1.2.0
xlrd==2.0.1
xlwt==1.3.0
